+++ src/components/Heading/headingChangelogData.ts
@@ -10,6 +10,23 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Functionality"],
+    notes: [
+      "Replaces the error thrown with a console.warn.",
+      "Removes deprecated values for size and levels.",
+    ],
+  },
+  {
+    date: "2025-04-24",
+    version: "3.6.1",
+    type: "Update",
+    affects: ["Styles"],
+    notes: ["Increase the font weight for the `size` styles."],
+  },
+  {
     date: "2025-01-30",
     version: "3.5.3",
     type: "Update",