+++ src/components/Menu/menuChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2025-04-10",
+    version: "3.6.0",
+    type: "Update",
+    affects: ["Styles"],
+    notes: ["Sets the font weight for the `group` headings to `bold`."],
+  },
+  {
     date: "2025-02-13",
     version: "3.5.4",
     type: "Update",