+++ src/components/Notification/notificationChangelogData.ts
@@ -10,6 +10,27 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Functionality"],
+    notes: ["Removes deprecated `noMargin` prop."],
+  },
+  {
+    date: "2025-04-24",
+    version: "3.6.1",
+    type: "Update",
+    affects: ["Styles"],
+    notes: ["Variant font colors updated to sync with the VDL."],
+  },
+  {
+    date: "2025-04-10",
+    version: "3.6.0",
+    type: "Update",
+    affects: ["Styles"],
+    notes: ["Style updates to sync with the VDL."],
+  },
+  {
     date: "2024-03-14",
     version: "3.0.0",
     type: "Update",