+++ src/theme/components/progressIndicator.ts
@@ -6,9 +6,6 @@ import { createMultiStyleConfigHelpers } from "@chakra-ui/styled-system";
 import { StyleFunctionProps } from "@chakra-ui/system";
 
 interface ProgressIndicatorBaseStyle extends StyleFunctionProps {
-  // darkMode is a deprecated prop. It remains here so as not to
-  // cause a breaking change for those apps that still use it.
-  darkMode: boolean;
   size: ProgressIndicatorSizes;
   labelPlacement: ProgressIndicatorLabelPlacements;
 }
@@ -48,11 +45,10 @@ const getCircularContainerFlexDir = (labelPlacement) => {
 
 const ProgressIndicator = defineMultiStyleConfig({
   baseStyle: definePartsStyle(
-    ({ darkMode, size, labelPlacement }: ProgressIndicatorBaseStyle) => {
+    ({ size, labelPlacement }: ProgressIndicatorBaseStyle) => {
+      const circularLabelMargin = size === "default" ? "xs" : "xxs";
       return {
-        color: darkMode
-          ? "dark.ui.typography.heading"
-          : "ui.typography.heading",
+        color: "ui.typography.heading",
         fontSize: "desktop.body.body2",
         _dark: {
           color: "dark.ui.typography.heading",
@@ -66,13 +62,13 @@ const ProgressIndicator = defineMultiStyleConfig({
             display: "block",
             circle: {
               _first: {
-                stroke: darkMode ? "ui.gray.dark" : "ui.gray.light-cool",
+                stroke: "ui.gray.light-cool",
                 _dark: {
                   stroke: "dark.ui.bg.hover",
                 },
               },
               _last: {
-                stroke: darkMode ? "ui.white" : "ui.link.primary",
+                stroke: "ui.link.primary",
                 _dark: { stroke: "dark.ui.link.primary" },
               },
             },
@@ -85,10 +81,10 @@ const ProgressIndicator = defineMultiStyleConfig({
           width: "fit-content",
         },
         circularLabel: {
-          marginBottom: labelPlacement === "top" ? "xxs" : 0,
-          marginLeft: labelPlacement === "right" ? "xxs" : 0,
-          marginRight: labelPlacement === "left" ? "xxs" : 0,
-          marginTop: labelPlacement === "bottom" ? "xxs" : 0,
+          marginBottom: labelPlacement === "top" ? circularLabelMargin : 0,
+          marginLeft: labelPlacement === "right" ? circularLabelMargin : 0,
+          marginRight: labelPlacement === "left" ? circularLabelMargin : 0,
+          marginTop: labelPlacement === "bottom" ? circularLabelMargin : 0,
           fontSize:
             size === "default"
               ? "desktop.label.label1"
@@ -98,13 +94,13 @@ const ProgressIndicator = defineMultiStyleConfig({
           // Hard to target this specific element without using
           // "Progress" as the key name in index.ts
           "> div": {
-            bg: darkMode ? "ui.white" : "ui.link.primary",
+            bg: "ui.link.primary",
             _dark: {
               bg: "dark.ui.link.primary",
             },
           },
           flex: 25,
-          bg: darkMode ? "ui.gray.dark" : "ui.gray.light-cool",
+          bg: "ui.gray.light-cool",
           height: {
             base: "4px",
             md: size === "default" ? "8px" : "4px",